+++ source/adios2/toolkit/format/bp5/BP5Deserializer.h
@@ -44,7 +44,7 @@ public:
     {
         size_t Timestep;
         size_t WriterRank;
-        size_t StartOffset;
+        uint64_t StartOffset;
         size_t ReadLength;
         char *DestinationAddr;
         bool DirectToAppMemory;
